pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "deckflow"
version = "0.1.3"
description = "Programmatic PowerPoint manager"
authors = [{ name="Nicolas Pogeant" }]
readme = "README.md"
requires-python = ">=3.9" # Support 3.9+
dependencies = [
    "python-pptx>=1.0.2",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
]

[project.urls]
Homepage = "https://github.com/npogeant/deckflow"
